api/requirements.txt
addict==2.4.0
ag-ui-protocol==0.1.8
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
argcomplete==3.6.2
attrs==25.3.0
blis==1.3.0
boto3==1.40.21
botocore==1.40.21
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.22.0
cohere==5.17.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.11
defusedxml==0.7.1
distro==1.9.0
easyocr==1.7.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
eval_type_backport==0.2.2
fastapi==0.116.1
fastavro==1.12.0
filelock==3.19.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.7.0
genai-prices==0.0.24
google-auth==2.40.3
google-genai==1.32.0
greenlet==3.2.4
griffe==1.13.0
groq==0.31.0
groundingdino @ git+https://github.com/IDEA-Research/GroundingDINO.git@856dde20aee659246248e20734ef9ba5214f5e44
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
invoke==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
lockfile==0.12.2
logfire-api==4.3.6
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mcp==1.13.1
mdurl==0.1.2
mistralai==1.9.9
mpmath==1.3.0
multidict==6.6.4
murmurhash==1.0.13
networkx==3.5
nexus-rpc==1.1.0
ninja==1.13.0
numpy==2.2.6
openai==1.102.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opentelemetry-api==1.36.0
packaging==25.0
pid==3.0.4
pillow==11.3.0
platformdirs==4.4.0
playwright==1.55.0
polars==1.32.3
preshed==3.0.10
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycocotools==2.0.10
pydantic==2.11.7
pydantic-ai==0.8.1
pydantic-ai-slim==0.8.1
pydantic-evals==0.8.1
pydantic-graph==0.8.1
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.2.3
pyperclip==1.9.0
python-bidi==0.6.6
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
pyzbar==0.1.9
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.13.1
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.1
segment-anything==1.0
service==0.6.0
setproctitle==1.3.6
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sse-starlette==3.0.2
starlette==0.47.3
supervision==0.26.1
sympy==1.14.0
temporalio==1.16.0
tenacity==9.1.2
thinc==8.3.6
tifffile==2025.8.28
timm==1.0.19
tokenizers==0.22.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.0
typer==0.17.3
types-protobuf==6.30.2.20250822
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.15.0
ultralytics==8.3.189
ultralytics-thop==2.0.16
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.3
yapf==0.43.0
yarl==1.20.1
zipp==3.23.0
